+++ config_modules_vmware/controllers/vcenter/dvs_network_io_control_policy.py
@@ -8,6 +8,7 @@ from typing import Tuple
 from pyVmomi import vim  # pylint: disable=E0401
 
 from config_modules_vmware.controllers.base_controller import BaseController
+from config_modules_vmware.controllers.vcenter.utils.vc_dvs_utils import is_host_disconnect_exception
 from config_modules_vmware.framework.auth.contexts.base_context import BaseContext
 from config_modules_vmware.framework.auth.contexts.vc_context import VcenterContext
 from config_modules_vmware.framework.clients.common import consts
@@ -25,6 +26,7 @@ SWITCH_NAME = "switch_name"
 GLOBAL = "__GLOBAL__"
 OVERRIDES = "__OVERRIDES__"
 OFFLOAD_NONE = "None"
+IGNORE_DISCONNECTED_HOSTS = "ignore_disconnected_hosts"
 
 
 class DVSNetworkIOControlPolicy(BaseController):
@@ -107,7 +109,8 @@ class DVSNetworkIOControlPolicy(BaseController):
                   "switch_name": "Switch-A",
                   "network_io_control_status": true
                 }
-              ]
+              ],
+              "ignore_disconnected_hosts": true
             }
 
         :param context: Product context instance.
@@ -160,7 +163,8 @@ class DVSNetworkIOControlPolicy(BaseController):
                   "switch_name": "Switch-A",
                   "network_io_control_status": true
                 }
-              ]
+              ],
+              "ignore_disconnected_hosts": true
             }
 
         :param vc_vmomi_client: VC vmomi client instance.
@@ -175,6 +179,7 @@ class DVSNetworkIOControlPolicy(BaseController):
         current = []
         desired_global_network_io_control_value = desired_values.get(GLOBAL, {}).get(DESIRED_KEY)
         overrides = desired_values.get(OVERRIDES, [])
+        ignore_disconnected_hosts = desired_values.get(IGNORE_DISCONNECTED_HOSTS, False)
         # desired_network_io_control_value = desired_values.get(DESIRED_KEY)
         try:
             all_switch_refs = vc_vmomi_client.get_objects_by_vimtype(vim.DistributedVirtualSwitch)
@@ -219,6 +224,14 @@ class DVSNetworkIOControlPolicy(BaseController):
                         dvs_ref.EnableNetworkResourceManagement(desired_network_io_control_value)
                         previous.append({SWITCH_NAME: dvs_ref.name, DESIRED_KEY: current_network_io_control_value})
                         current.append({SWITCH_NAME: dvs_ref.name, DESIRED_KEY: desired_network_io_control_value})
+                    except vim.fault.DvsOperationBulkFault as e:
+                        if is_host_disconnect_exception(e) and ignore_disconnected_hosts:
+                            previous.append({SWITCH_NAME: dvs_ref.name, DESIRED_KEY: current_network_io_control_value})
+                            current.append({SWITCH_NAME: dvs_ref.name, DESIRED_KEY: desired_network_io_control_value})
+                            logger.info(f"Ignore disconnected hosts caused exception - {e}")
+                        else:
+                            logger.exception(f"An error occurred: {e}")
+                            errors.append(str(e))
                     except Exception as e:
                         logger.exception(f"An error occurred: {e}")
                         errors.append(str(e))